projects/vibecheck-flask/requirements.txt
alembic==1.4.3
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
bidict==0.21.2
black==20.8b1
blinker==1.4
boto3==1.34.58
botocore==1.34.58
certifi==2024.2.2
cffi==1.16.0
cfgv==3.3.1
chardet==4.0.0
click==7.1.2
distlib==0.3.8
filelock==3.13.1
flake8==3.9.2
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==4.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gunicorn==20.1.0
identify==2.2.15
idna==2.10
iniconfig==2.0.0
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==1.0.1
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==24.0
pathspec==0.8.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pytest==8.1.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==6.0.1
regex==2020.11.13
requests==2.25.1
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==1.3.22
toml==0.10.2
typed-ast==1.5.5
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.25.1
Werkzeug==2.0.3
wrapt==1.12.1
WTForms==2.3.3
